Well, well well. VBO Patch da SBF is working
The UI is terrible of course. Only a few strange people like myself can write JSONPATCH commands so other than training and documenting we need also an option for a better UI, where people can, e.g Pick from an existing Object a Value? and then decide the return value. Also add the JMESPATH (e.g as:image.*. to multiple JSON Pointers automatic casting. That needs to happen per Object

Means, the user adds the invalid JSONPointer in the single JSONPATCH operation. We validate (as true), then per Object we get the result for that single JMESPATH, If there is a match, we rewrite on the fly the dynamic operation into multiple JSONPOINTERS ones). I feel like a wizard on this and know i will not be remembered by this but for my curry!