Big push after local refactor
This includes the new Entities, the base plugins, the main delooper service, the depencies on Google Client and routing. This is not 1:1 to the working version i have locally but its better to push this branch first so it does not appear magically as a whole block in Drupal. This has pieces also that are in refactoring process!