Move validation to its own method. Add strict mode, fix invalid key access
This adds a protected validateAmiSet() function that takes the Set Config ($data), the CSV data ($file_data_all) and a boolean and validates either normal (for required elements) of strict (for all column keys defined in the original source data (means 1:1 with the reference CSV). Strict mode is needed for Canned Ingests (more later on that, i'm just a human!)

Also on the last iteration (where we order ingests based on parents first, then children) i was still allowing non existing ROW numbers to be referenced, this should be fixed now

And now we return the Invalid Rows (no message why yet). This is all on the preprocessor of course. Since the actual ADO generation and file upload will have its own errors if that fails for other reasons. Pre process does a lot but not everything or it would not be efficient at all.

How to test?

- Create an AMI set
- Download the resulting CSV
- Mess it up by making one parent invalid (non existing row, non existing UUID), see the errors, will process but will tell you which Row failed.
- Mess it up again, change the headers and put invalid fields there. Try again, you will get invalid message and processing won't happen
- Mess it up again, put on every parent ROW the number 10000.  Ingest won't happen.

Hope this makes things better. Still need to work on the actual CSV report. but that is another issue.

PS: Validating this is a lot of effort and we may want to be even stricter!