Update amiSetEntityDeleteProcessedForm.php
Delete will attempt to delete but will first validate per UUID access. So no "pre" to batch error messages.
But if no valid UUIDs could be fetched then we will just tell the user " A general " why instead of exposing "security" and  "data" to the user. This is on purpose, please do not ask for label of Which Bundle ADO could not be deleted.