Update amiSetEntityProcessForm.php
First pass on Processing (no Update/patch yet).

Checks if the bundles passed via the config are still valid and the user can actually create ADOs for those if a single one fails no processing will be allowed at all. Same as with Delete ADOs, we do not expose data to the user.

This will get a few more options when Update/Patch are done