First pass on I7 importer
Still work to do and clean ups.
Batch is not triggering at the end because the Form is Ajax Driven. I need to change the button at the end to a normal Submit (no ajax) when the plugin annotation batch == true. That should also solve the issue i have with not being redirected to the AMI set and having to "click" it.