Ok, batch is running!
well, well, well. Now some cleanup and adding the offset and max as Form inputs for the Solr Plugin,

Added a new CSV_append method that will use the first 100 rows fetched inside a CSV to keep appending. This is a bit risky if the fields change on each fetch. One way of avoiding the problem is to generate multiple CSVs and then iterate over each one getting its headers, normalize and join. I do not want to keep 300,000 Objects in an PHP array, just in case ....