Update SolrImporter.php
95% of this is $fom_state mangling and passing around.
The main idea is that the getInfo() method keeps setting
$form_state->setValue(['pluginconfig','ready'], FALSE); until all its needed data is fetched, including mappings, etc.
New is that if you put a 0 in the rows to fetch input it will use the actual value of the resulting Solr quey. Nice.

Batch operation increments correctly in 200 Documents per run and uses the offset too.

public static function finishfetchFromSolr() is still not used and i'm trying to figure out if i keep appending to the CSV or use in memory array to do so $context['results']['processed'][] ?