Got the Children offset right!
Also, Children are getting the right Data stream URLS and assigning to the right header/fields.
Still, the variable Header situation is not yet solved yet, requires probably passing also between batches the headers so each new sub batch knows what it may be missing and /or apply an offset based on the previous one to match the right column. 24/48 hrs more and we have a full blow Islandora importer and we are done with this.