Give Importer Plugins more power over the data
Adds providesKeys and providesTypes. That way Solr data that comes back is not used for UI but only the keys we generate. Makes all cleaner and faster.
Also some UI improvements. Children mappings are only shown IF collapse is disabled. @alliomeria i could not reproduce the false collapsing? So wonder if there is something else missing in your code? Let's try this tomorrow again