Does less manual cleaning and trusts fputcsv() more for escaping
@alliomeria this should fix the double encoding issue, but
just to be sure, could you try Solr import 7 with a collection that has complex weird characters in its descriptions/titles (maybe NYHS?)
and also with a google sheet using, e.g Japanese + "" and & and ' somewher?

WE can test tomorrow together. Thing is i removed a lot of code that i "for some reason" had added, so wonder if i will not break now something somewhere else, and thus "testing" is required!