New thing for me
Since we can not predict the final extension of a download upfront without download i check now using gob() for the first part of the possible filename.
If there is already a single (only a single one) File with the same starting future name (which is consistent since we use the URL of the file to generate it) then we reuse the already downloaded one. Will add also a button that deletes all files for an AMI set in the future and one that forces the download even if there.

Also this piece of code ensures ZIP files (After first upload) are moved into private://ami/zip once the AMI set is saved