C'mon ZipArchive WHY! You can not stream from remote
I have been willing to write the first PHP remote streamer library from forever and i never have the time. So for now, for the people that decide to put ALLL on S3 we have to download the file.. no other option.
So we inject the StrawberryfieldFileMetadataService to reuse the ::ensureFileAvailability function.
But we can not delete the file here since we may need it for the rest of our times.