Offset ::csv_count(File $file) by 1 and get around PHP8 bug
PHP 8 fails seeking to a certain like (like in this case counting) because the flags passed are not taken in account. This was tested in PHP 8.0.16. To avoid writing double code i decided to take a naive approach and actually us fgetcsv() without return to move through the file since that function has no issues with our CSVs and JSON added data in cells (weird that seek has). It works, tested on php 7.4 and 8.0.16