Adds the data VBO View Event Subscriber
This event subscriber will copy URL arguments just before a Batch operation gets started. Has no idea if they will be used or not (that is on the Facet processor). Avoids overwriting them with empties once the Batch starts. Since this event is triggered when a change on the VBO form happens, we do not need to clear it up at all. Will be fresh with the needed values every time.