Reverting this
Need to ask @alliomeria / @karomabiles about what to expect when "updating" a published entity and passing a new "draft state". Because what will happen now (which is correct but annoying bc of workflows) is that the published one will stay as it  was and a new revision (draft) will be added but not enabled yet. Which can be then "made" published via Content/publish action. But it might need documentation?