Much better increment
Small change. We won't longer use the $per_page to calculate the children max. We set the max to static increment and use the per_page only to count the top objects. this makes the 75% reducing factor when requesting all 500 and getting less top ones to only make the top objects less still allowing the children to increment more. In practice this keeps numbers low but does reduce less when there are many children making the fetch faster