Fixes the problem with Child CMODEL Mappings v/s Top Object ones
At least this way @alliomeria and @karomabiles your decisions stick and are brought from child to top and back. Also changes a bit the messages we send during the batch fetching from Solr.
But we need to test this, i kinda feel now things are a bit slower than before. Maybe it is me?