Preserve UUID when reusing a file
We try hard to reuse files. This of course opens the question of. AMI users are power users, still should be allow file reuse without checking access? Or said differently, we can reuse, but maybe not change the owner? Still this fixes it and TBH this only will trigger if the source/destination are the same, means files coming from S3.