Double check if the cached File is still there
Because how Drupal's cron work, a file might be temporary after deleting a Set ingested object, our cache still in place but the actual File gone because of Drupal.... This forces reprocessing even if Cache still exists in the absence of a real file. This is not new. It would fail anyways on ingest (up in the main queue worker) but without the chance of getting it again .. this at least will try