refactor \Drupal\ami\AmiUtilityService::preprocessAmiSet and small File Download fix
 A lot is new here:
- First we preprocess relationship values by trying to "array-ify" them always, either if its a single value, a JSON or a ; separated string.
Then we do as we do normally on each parent (multiple_ traversing the tree up but just more times. But once that is done things get a bit more messy and we use a different sorting (naive but works) by putting Parents first but If a parent has a child a parent added already before, we don't append, we unshift it trying to ensure this way that parents that have children that are also parents get ingested sooner. If not it will still work (ha!) but there will be retries and that makes the queue slower.