Improvements and cleanups related to multi rels on queue worker
Most of the code was there already but i found small other uses cases and neededs related to files! This now cleans up files separated by ; better, and avoids also downloading FOLDERS!
Also sometimes the ingest will fail NOT because the bundle is not present but because a New Type was added that has no mapping. So we add this here too as a message