Remove spaces and dots from the beginning in any combination
Regular expression. Will clean anything that is spaces and dots and the combinatory. To be extra safe, after cleaning we check if anything is left and if not we actually default the the URL given name...

I only do this for Names provided by a header/remove HTTP idea. If a user decides to use local filenames or S3 ones, we don't want to modify their names at all (like really)