Move remote mime exceptions to just after we fetch them. Matches SBF ISSUE
https://github.com/esmero/strawberryfield/issues/274

This will "correct" a few hardcoded ones from remote Fedora/Islandora datastreams and map them back to what we know they should. In the future we can add a config for this. Now the "calculated from the remote extension" + extension calculated from the wrong (and corrected) mimetype will match and we will be able to use either JPEG or JPG depending on what is coming preserving OBJ/Material needs for 3D