The issue with View ID + Display ID length can trigger everywhere
This targets the issue for our VBO collections also

That said, it is futile! Views Operation Batch (that is where i came up with the name in the first place ...) uses this
https://git.drupalcode.org/project/views_bulk_operations/-/blob/4.2.x/src/ViewsBulkOperationsBatch.php#L76

Just tested with very long (but valid machine names) and Drupal basically dies. SO NOT ON US!! Lesson. Keep your Views/Display names short ok?

@alliomeria gosh...