Deals with moving target for out VBO Event subscriber
Since Drupal 10.2 we can't no longer just relay on a 'op' in the input to avoid saving. This here attempts first to check if the VBO operation is happening under the actual Views ROUTE (so the page that the views handles) OR, if someone is using layout builder if 'op' != 'do' (might be another moving target) and there is no batch id.