make sure actions (when running on hydroponics) run as the user that sent them
Why?

Actions don't have enough logic or data to switch users, so we do that for them, also blocking Anonymous users to even attempting it. We do check for action access though everytime, even before this, but without this an AMI set created via an AMI CSV export action ended belonging to the anonymous user, which is totally terrible @alliomeria